Spe 157 implement o auth (#72)
* get to login google page but bad redirect after login

* more changes

* 3 changes

* tried with github

* Get attrs and make response body but trouble with transactions

* Added oauth to frontend without photos

* Moved Google oauth verification to backend and removed needless files

* Fixed frontend a bit

* Fixed env var and some backend constraints

* Added photos to OAuth

* Merged, but delivery not working

* Adjusted frontend

* Added temporary page reload

* Removed tracing

* Fixed cancel bug

* Fixed some bugs

* Fixed google auth

+++ src/main/java/org/eventhub/main/service/impl/PhotoServiceImpl.java
@@ -91,10 +91,25 @@ public class PhotoServiceImpl implements PhotoService {
     public void deleteProfileImage(UUID ownerId, UUID imageId) {
 
         String blobName = this.readById(imageId).getPhotoName();
-        BlockBlobClient blockBlobClient = this.blobContainerClientUser.getBlobClient(blobName).getBlockBlobClient();
-        blockBlobClient.delete();
+        if (!blobName.startsWith("GPhoto")) {
+            BlockBlobClient blockBlobClient = this.blobContainerClientUser.getBlobClient(blobName).getBlockBlobClient();
+            blockBlobClient.delete();
+        }
         userService.deleteImage(ownerId, this.readByIdEntity(imageId));
         photoRepository.delete(readByIdEntity(imageId));
+    }
+
+    @Override
+    public PhotoResponse addUserPhotoByUrl(UUID userId, String photoUrl) {
+        Photo photo = new Photo();
+        photo.setId(UUID.randomUUID());
+        photo.setPhotoName(String.format("GPhoto%s", photo.getId()));
+        photo.setPhotoUrl(photoUrl);
+
+        userService.addImage(userId, photo);
+        PhotoResponse response = photoMapper.entityToResponse(photoRepository.save(photo));
+
+        return response;
     }
 
     @Override
